fix(cmp): add vscode cmp menu icons, menu direction, fix snippet engine
vscode menu icons just make the completion menu look better. ive tried the other menu view types too
and prefer the vertical version, but now have the completion menu showing suggestions closest to
cursor based on if it opens above or below. snippet engine was simply just required to install the
snippet plugin and specify it in the config as i thought

fix #5