feat(snippets): more snippets including first vim regex for capitalising first letter (#24)
also adding reloading of nvim config without having to quit vim so that i can play with things more
easily. mappings don't seem to be updated but mostly everything else comes in. interestingly trying
regex within the snippy text transforms it actually takes vim regex patterns, see the manual for
more info on this as that was very helpful and not documented anywhere else that i searched

re #16